pbot: Check retry also for forced builds.
The reason for not checking it before was that that made it impossible to
force a re-run of already run benchmarks for a given runset.  Now the retry
logic will detect that those benchmarks have been run and won't run them
again.

The proper way to solve this is to get the information on which benchmarks
to run not out of the build logs but out of the database.  If you want benchmarks
re-run, delete them from the DB and force the runset again.